create EditorInterfaces.ts
